# RuntimeError: asyncio.run() cannot be called from a running event loop