--- a/.devcontainer/devcontainer.json
+++ b/.devcontainer/devcontainer.json
@@ -3,8 +3,7 @@
   // TODO: @alexanderilyin switch to SemVer in image tags
   "image": "ghcr.io/partcad/devcontainer-partcad:latest",
   "postCreateCommand": {
-    // TODO: @alexanderilyin: sudo apt-get install file
-    // TODO: @alexanderilyin: sudo apt-get install xxd
+    "setup-utils": "sudo apt-get update && sudo apt-get install -y file xxd",
     "pre-commit": "pre-commit install --config .devcontainer/.pre-commit-config.yaml"
   },
   // "mounts": [