--- a/.github/workflows/python-build.yml
+++ b/.github/workflows/python-build.yml
@@ -24,7 +24,7 @@ jobs:
         # TODO(clairbee): Add pypy versions to the matrix?
         python-version: ["3.10", "3.12"]
         #python-version: ["3.10", "3.11", "3.12"]
-        # TODO(clairbee): prescrive specific versions since nlopt often causes problems on Windows or MacOS?
+        # TODO(clairbee): prescribe specific versions since nlopt often causes problems on Windows or MacOS?
         # python-version: ["3.10.11", "3.11.8", "3.12.3"]
         exclude:
           # Reduce costs by testing one version of Python on MacOS