--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -39,7 +39,6 @@
       ]
     }
   ],
-
   "files.trimFinalNewlines": true,
   "files.encoding": "utf8",
   "editor.useTabStops": false,
@@ -107,7 +106,10 @@
   "python.terminal.activateEnvInCurrentTerminal": true,
   "python.envFile": "${workspaceFolder}/.env",
   "cSpell.words": [
-    "coderabbitai"
+    "coderabbitai",
+    "gource",
+    "libx",
+    "ultrafast",
+    "vcodec"
   ],
-
 }