## Reproduce Scenario

```bash
poetry install
poetry shell
```

Follow [Tutorial — PartCAD Documentation](https://partcad.readthedocs.io/en/latest/tutorial.html#export-the-part)

```bash
pc init
cat <<EOF > partcad.yaml
import:
  # Public PartCAD repository (reference it explicitly if required)
  pub:
    type: git
    url: https://github.com/partcad/partcad-index.git
EOF
pc list
echo "translate (v= [0,0,0])  cube (size = 10);" > test.scad
pc add-part scad test.scad
pc inspect :test
pc render -t stl :test
```

## BEFORE

```
⬢ [Docker] ❯ pc -vvv render -t stl :test
INFO:  DONE: InitCtx: /tmp: 0.00s
INFO:  DONE: Render: this: 0.00s
ERROR: PartCAD CLI exception
Traceback (most recent call last):
  File "/workspaces/partcad/partcad-cli/src/partcad_cli/cli.py", line 216, in main
    cli_render(args, ctx)
  File "/workspaces/partcad/partcad-cli/src/partcad_cli/cli_render.py", line 168, in cli_render
    asyncio.run(cli_render_async(args, ctx, packages))
  File "/usr/local/python/3.10.16/lib/python3.10/asyncio/runners.py", line 44, in run
    return loop.run_until_complete(main)
  File "/usr/local/python/3.10.16/lib/python3.10/asyncio/base_events.py", line 649, in run_until_complete
    return future.result()
  File "/workspaces/partcad/partcad-cli/src/partcad_cli/cli_render.py", line 140, in cli_render_async
    prj.render(
  File "/workspaces/partcad/partcad/src/partcad/project.py", line 1460, in render
    asyncio.run(self.render_async(sketches, interfaces, parts, assemblies, format, output_dir))
  File "/usr/local/python/3.10.16/lib/python3.10/asyncio/runners.py", line 33, in run
    raise RuntimeError(
RuntimeError: asyncio.run() cannot be called from a running event loop
sys:1: RuntimeWarning: coroutine 'Project.render_async' was never awaited
RuntimeWarning: Enable tracemalloc to get the object allocation traceback
```

## AFTER

```
⬢ [Docker] ❯ pc -vvv render -t stl :test
INFO:  DONE: InitCtx: /tmp: 0.00s
DEBUG: Rendering: /tmp/./test.stl
DEBUG: test: WARNING: CPU random generator seem to be failing, disabling hardware random number generation
DEBUG: test: WARNING: RDRND generated: 0xffffffff 0xffffffff 0xffffffff 0xffffffff
DEBUG: test: Geometries in cache: 2
DEBUG: test: Geometry cache size in bytes: 1456
DEBUG: test: CGAL Polyhedrons in cache: 0
DEBUG: test: CGAL cache size in bytes: 0
DEBUG: test: Total rendering time: 0:00:00.001
DEBUG: test:    Top level object is a 3D object:
DEBUG: test:    Facets:          6
DEBUG: test: 
INFO:  DONE: Render: this: 0.19s
```

## Copilot Summary


This pull request includes several changes to improve code quality, add new features, and update configurations. The most important changes include the addition of Gource visualization instructions, updates to `.vscode` settings, and various code quality improvements in `partcad/src/partcad/project.py`.

### Documentation Update:
* Added instructions for installing and running Gource, a software version control visualization tool, in `mkdocs/docs/development.md`.

### Configuration Updates:
* Added new words to the `cSpell.words` list in `.vscode/settings.json` to improve spell-checking accuracy.
* Enabled trimming of final newlines in files and set file encoding to UTF-8 in `.vscode/settings.json`.

### Code Quality Improvements:
* Fixed various typos and improved code readability in `partcad/src/partcad/project.py`. [[1]](diffhunk://#diff-c089aaff8ced795345ac59e0c82517e60e0934843eadb83fd1aac8631c637ba6L268-R268) [[2]](diffhunk://#diff-c089aaff8ced795345ac59e0c82517e60e0934843eadb83fd1aac8631c637ba6L895-R896) [[3]](diffhunk://#diff-c089aaff8ced795345ac59e0c82517e60e0934843eadb83fd1aac8631c637ba6L1047-R1047)
* Changed the `render` method to be asynchronous in `partcad/src/partcad/project.py` to improve performance. [[1]](diffhunk://#diff-c089aaff8ced795345ac59e0c82517e60e0934843eadb83fd1aac8631c637ba6L1451-R1451) [[2]](diffhunk://#diff-c089aaff8ced795345ac59e0c82517e60e0934843eadb83fd1aac8631c637ba6L1460-R1460)

### DevContainer Configuration:
* Added TODO comments for installing `file` and `xxd` packages in `.devcontainer/devcontainer.json` for future setup improvements.

<!--

CodeRabbit will automatically analyze your PR and provide:
- Code review comments
- Security analysis
- Performance insights
No action needed in this section - let AI do the heavy lifting!

@coderabbitai
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
	- Added a new section in the documentation for "Gource", detailing its installation and usage for visualizing project history.
- **Documentation**
	- Included commands for installing Gource and generating a video representation of project activity.
- **Chores**
	- Enhanced spell-checking in the editor by adding specific terms to the configuration.
- **Refactor**
	- Updated the `render` method in the `Project` class to be asynchronous, improving efficiency in handling concurrent rendering tasks.
- **Bug Fixes**
	- Corrected the spelling of "neccessary" to "necessary" in the help text for the `-p` argument.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->